flags functionality changed
